requirements.txt
telethon>=1.36,<2.0
python-dotenv>=1.0
simple-term-menu>=1.6
